ref(prettier): extract tslint rule into own file
