+++ src/prettier/index.ts
@@ -7,6 +7,7 @@ import {
   url
 } from '@angular-devkit/schematics';
 import { installDependencies, PackageJson } from '../lib';
+import { patchTsLintConfiguration } from './rules';
 
 export default function(): Rule {
   return chain([
@@ -30,25 +31,5 @@ function registerPrettier(): Rule {
     tree.overwrite('package.json', packageJson.stringify());
 
     return mergeWith(apply(url('./templates'), []));
-  };
-}
-
-function patchTsLintConfiguration(): Rule {
-  return (tree: Tree) => {
-    const tslintFile = tree.read('tslint.json');
-
-    if (!tslintFile) {
-      return;
-    }
-
-    const tslintJson = JSON.parse(tslintFile!.toString('utf-8'));
-
-    if (Array.isArray(tslintJson.extends)) {
-      tslintJson.extends.push('tslint-config-prettier');
-    } else {
-      tslintJson.extends = [tslintJson.extends, 'tslint-config-prettier'];
-    }
-
-    tree.overwrite('tslint.json', JSON.stringify(tslintJson, null, 2));
   };
 }
