+++ src/prettier/rules/index.ts
@@ -1,0 +1,1 @@
+export * from './patch-ts-lint-configuration';
